+++ examples/client-example/src/App.js
@@ -1,6 +1,5 @@
 import Knock from "@knocklabs/client";
-import { useEffect, useMemo } from "react";
-import create from "zustand";
+import { useEffect, useMemo, useState } from "react";
 
 import "./App.css";
 
@@ -17,10 +16,10 @@ const useNotificationFeed = (knockClient, feedId) => {
       auto_manage_socket_connection: true,
       auto_manage_socket_connection_delay: 500,
     });
-    const notificationStore = create(notificationFeed.store);
+
     notificationFeed.fetch();
 
-    return [notificationFeed, notificationStore];
+    return [notificationFeed, notificationFeed.store];
   }, [knockClient, feedId]);
 };
 
@@ -29,6 +28,7 @@ function App() {
     knockClient,
     process.env.REACT_APP_KNOCK_CHANNEL_ID,
   );
+  const [feedState, setFeedState] = useState(feedStore.getState());
 
   useEffect(() => {
     knockClient.preferences
@@ -40,6 +40,20 @@ function App() {
         console.error(e);
       });
   }, []);
+
+  // Consume the store
+  useEffect(() => {
+    // What to do on updates
+    const render = (state) => {
+      setFeedState(state);
+    };
+
+    // What to do on initial load
+    render(feedStore.getInitialState());
+
+    // Subscribe to updates
+    feedStore.subscribe(render);
+  }, [feedStore]);
 
   useEffect(() => {
     const teardown = feedClient.listenForUpdates();
@@ -59,7 +73,7 @@ function App() {
     return () => teardown?.();
   }, [feedClient]);
 
-  const { loading, items, pageInfo } = feedStore((state) => state);
+  const { loading, items, pageInfo } = feedState;
 
   return (
     <div className="App">
@@ -71,6 +85,8 @@ function App() {
       {items.map((item) => (
         <div key={item.id} className="feed-item">
           ID: {item.id}
+          <br />
+          Has been read: {item.read_at ? "true" : "false"}
           <br />
           Actor ID: {item.actors?.[0]?.id}
           <br />