+++ packages/react-core/src/modules/feed/hooks/index.ts
@@ -3,4 +3,5 @@ export { default as useFeedSettings } from "./useFeedSettings";
 export {
   default as useNotificationStore,
   useCreateNotificationStore,
+  type Selector,
 } from "./useNotificationStore";