+++ packages/expo/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @knocklabs/expo
 
+## 0.3.0
+
+### Minor Changes
+
+- 8ba5dcb: [JS] Support React 19 in React SDKs
+
+### Patch Changes
+
+- Updated dependencies [8ba5dcb]
+  - @knocklabs/client@0.12.0
+  - @knocklabs/react-core@0.5.0
+  - @knocklabs/react-native@0.6.0
+
 ## 0.2.12
 
 ### Patch Changes